:rocket: [Update] Separated Train and model modules
