+++ train.py
@@ -17,21 +17,8 @@ from dotenv import load_dotenv
 load_dotenv()
 
 from dataset import SpeechDataModule
-from Conformer import ConformerEncoder, LSTMDecoder
+from model import ConformerASR
 from utils import GreedyDecoder
-
-
-class ConformerASR(nn.Module):
-    def __init__(self, encoder_params, decoder_params):
-        super(ConformerASR, self).__init__()
-        self.encoder = ConformerEncoder(**encoder_params)
-        self.decoder = LSTMDecoder(**decoder_params)
-
-    def forward(self, x):
-        encoder_output = self.encoder(x)
-        decoder_output = self.decoder(encoder_output)
-        return decoder_output
-
 
 class ASRTrainer(pl.LightningModule):
     def __init__(self, model, args):
